Tempus: Add Observers to Steppers
Added observers to BackwardEuler, ForwardEuler, ExplicitRK, DIRK,
and IMEX-RK steppers.  This is an initial guess at the needed observer
functions for each stepper.  We will adjust as needs become clearer.
The SolutionHistory and the Stepper are passed to the observers, so
that application developers can have access to the solution, stepper
member data, and the ModelEvaluators.

Misc. changes:
 * Tested changes on Mac OS and linux (ceerws1113), and all tests
 * passed.